MIRTH-4  Transformer mapping step names should reflect the mapping variable
OTHER BUGS FIXED IN PROCESS:
Fixed a bug that caused the client to not allow saving of a channel after a transformer or filter was imported.
Fixed a bug that allowed blank variables to be entered int he transformer panel.
Fixed a bug that allowed blank variables to be entered by adding a new step.
Changed a lot of logic in the code to prevent future bugs.


git-svn-id: https://svn.mirthcorp.com/connect/trunk@1006 c801cdd1-710f-0410-a44a-8d9edceeb29b