MIRTH-1479 - Changed XmlUtil to have all static methods rather than using getInstance().  Added the old Entities class back and deprecated it.
MIRTH-1528 and MIRTH-1490 - Added the ChannelController getChannelName and getChannelId methods back and deprecated them.

MIRTH-1536 - Removed extra channel name and id references from the ReferenceListFactory.  They were in both the message category and the new channel category.

Added string replacement step with version checking to the ImportConverter.  This is not currently being used.

git-svn-id: https://svn.mirthcorp.com/connect/trunk@5091 c801cdd1-710f-0410-a44a-8d9edceeb29b