Removed queued stat from channel statistics table and updated related queries. Removed queued stat from transactional statistics. The queued value in the dashboard table now comes directly from the ConnectorMessageQueue objects of a channel.
Removed method "removeMessagesFromQueues" method and replaced it with "invalidateQueues" which clears out the entire queue buffer, forcing a database query the next time data is requested from the queue. This helps to keep the size accurate, and keeping the queue from becoming out of sync with the actual data in the database.
Only poll the queue after processing if the message at the front of the queue has the same message and metadata ids as the message that was just processed.

Data sources for the queues are now set when the channel is deployed, rather than when the channel is started.
The queues are now invalidated when the channel is started.

If a runtime exception occurs during destination queue processing, invalidate the queue in case the message was deleted and a foreign key constraint was triggered. Otherwise the same message will be sent indefinitely.

Updated message/connector message delete logic to use CASCADE instead of deleting from each table.
Removed queries that are unnecessary now that CASCADE is used.

Remove All Messages now only works for channels that are stopped.
Remove Results and Remove Message will only delete processed messages while the channel is running. It will delete unprocessed (unfinished) messages if the channel is stopped.
Channels cannot be started while a delete is in progress. This is to ensure the message currently being processed is never deleted.
Remove Results now deletes in batches. The exact number per batch will be tuned in the future.

Modified code using removeConnectorMessages on the Client to use removeMessages instead. This is necessary in order to check if the message has been processed.
Removed the "removeConnectorMessages" method from the Client because it is no longer being used.
Issue: MIRTH-2250


git-svn-id: https://svn.mirthcorp.com/connect/trunk@5901 c801cdd1-710f-0410-a44a-8d9edceeb29b