MIRTH-2265: Implemented a new message pruner that is able to prune content separately from the entire message. Added junit tests to test the new pruner.
MIRTH-1699: Added the ability to provide the message pruner with a message archiver that will archive messages prior to pruning them. We still need to implement a message archiver.

MIRTH-2250: Removed the batch pruning option with the new message pruner since batch pruning is no longer possible with the new messaging engine.

git-svn-id: https://svn.mirthcorp.com/connect/trunk@5914 c801cdd1-710f-0410-a44a-8d9edceeb29b