MIRTH-1696: Updated the import result pop up window to display the number of messages successfully imported and the numbers of messages that failed to import due to error(s).
MIRTH-2250:

Added import_channel_id to the message (d_m) table to keep track of the original channel id for imported messages.

Modified the message import behavior so that the original message id is stored as the "import id" and a new message id is always generated on import.

Added support for importing messages from a zip file.

git-svn-id: https://svn.mirthcorp.com/connect/trunk@5940 c801cdd1-710f-0410-a44a-8d9edceeb29b