Cleaned up a TODO in DICOMUtil.
run(String arg) is required to initialize the ij.plugin.DICOM object. The argument serves multiple purposes. If it is null or empty, it opens a dialog to select a dicom file. If it is a path, it opens the specified image. Otherwise, it is the title of the dialog if dicom.show() is called.
Issue: MIRTH-2250


git-svn-id: https://svn.mirthcorp.com/connect/trunk@5958 c801cdd1-710f-0410-a44a-8d9edceeb29b