More improvements to DICOM memory performance.
The DicomOutputStream now writes to a dummy ByteCounterOutputStream that simply counts the number of bytes to be written. That number is then used as the buffer size of the ByteArrayOutputStream that the data is actually written to. This optimizes the memory requirement of the DicomObject to byte array method to 2x the size of the DicomObject. There should be no better performance possible of this task using these Dcm4che methods.

The amount of memory required to process a dicom file with no transformations in the latest build is slightly greater than 3x the size of the dicom file itself. It may be possible to improve this even further down to the lower limit of 2x the file size, but that remains to be seen. Still, there should be a large improvement over 2.2.x
Issue: MIRTH-1573


git-svn-id: https://svn.mirthcorp.com/connect/trunk@5999 c801cdd1-710f-0410-a44a-8d9edceeb29b