Even more improvements to DICOM memory usage.
Attachments are now chunked in 10,000,000 byte segments when they are inserted into the database. They are combined on retrieval. This dramatically improves the amount of memory required when retrieving them from the database.
Added segmentID and size to the attachments table.
Moved attachment selection queries from DonkeyMessageController to DonkeyDao, in order to perform the combining.

ASCII attachments are now always converted to byte arrays using UTF-8 when inserting into the database, instead of the system default charset. However, this process (specifically with UTF-8) requires an extremely large amount of memory relative to the size of the string. 
Created a new StringUtil.getBytesUncheckedChunked class/method which will convert strings to byte arrays in chunks, to alleviate the above issue. Currently set to 1,000,000 byte chunks.
When an ASCII attachment is reattached, it is combined using the UTF-8 charset. Then, if a different charset was specified, it is converted to a String using UTF-8, and then to the byte array of the specified charset. This process is skipped if UTF-8 is the specified charset.

Fix properties in RegexAttachmentHandler being cleared after the first message is processed.

Moved ByteCounterOutputStream from the server to Donkey so it can be used by getBytesUncheckedChunked.

Removed some orphaned references from default.xml

No longer clearing the template in FileDispatcher.
Removed some debug code from FileReceiver
Removed old Attachment class

MIRTH-2250
Issue: MIRTH-1573


git-svn-id: https://svn.mirthcorp.com/connect/trunk@6009 c801cdd1-710f-0410-a44a-8d9edceeb29b