Fix for 2.x
Fix memory leak that causes the dashboard to become unresponsive after the client has been active for a while (usually a day or two). The highlighters in the dashboard table were being added each time the table was updated, but never removed. 

Made the same fix to the channel table that ordinarily is fine, but could potentially have the same problem under certain user preferences.
Issue: MIRTH-2301


git-svn-id: https://svn.mirthcorp.com/connect/branches/2.x@6093 c801cdd1-710f-0410-a44a-8d9edceeb29b