Improved the JMS connector user-interface.
Replaced the "Preset" combo-box with a list of connection templates, in order to make the UI easier to understand.

The user is also prevented from overwriting connection templates that are pre-defined by Mirth (ActiveMQ, JBoss), to prevent confusion for them as well as the help desk.

Optimized the layout of the components.
Issue: MIRTH-1945


git-svn-id: https://svn.mirthcorp.com/connect/trunk@6186 c801cdd1-710f-0410-a44a-8d9edceeb29b