Added a "Deleted Connectors" row to the message browser advanced filter. It does what you would expect it to.
Cleaned up advanced filter text, alignment, and size.
ItemSelectionTable and TableModel now have the option to show the key column. Advanced Filter now shows the key column.

The search queries now use exclusion logic if the Deleted Connectors is selected, and inclusion logic if it is not selected.

Issue: MIRTH-2390


git-svn-id: https://svn.mirthcorp.com/connect/trunk@6195 c801cdd1-710f-0410-a44a-8d9edceeb29b