The HL7v2 data type can now accept strict XML data directly if the strict parser is enabled.
The data type plugins no longer require a method to check if the data type is Xml.

The message browser now checks the actual content to determine whether the content is Xml to enable/disable the formatting options, rather than relying on the plugin for that information.

Ran the code formatter.
Issue: MIRTH-2412


git-svn-id: https://svn.mirthcorp.com/connect/trunk@6217 c801cdd1-710f-0410-a44a-8d9edceeb29b