MIRTH-2426 Maps and Errors are now stored in the Content table (d_mc) rather than the Metadata table (d_mm) ConnectorMessage now holds a MapContent and ErrorContent instead of the data directly, this way the channel can know if it has already persisted either of them. ContentType now has 3 sub-types (Message, Map, Error). ContentType codes are now integers instead of characters.
MIRTH-2443
The encryptor has been moved out of the channel (and sub-objects) and into the JDBCDao directly. There is no need to hold both the unencrypted and encrypted content in memory during the duration of the message being processed. At the Dao level, the message is encrypted if necessary because storing into the DB. Encrypted messages are decrypted automatically at the Dao level as well.
PassthruEncryptor has been removed.
MessageContent now only stores a boolean of whether the content in it is encrypted. This more closely matches the content table.
Removed various encrypt/decrypt methods that are no longer necessary.

MIRTH-2442
Fixed encrypted queued messages not being able to be decrypted if channel encryption was disabled.

MIRTH-2464
DonkeyDao is now used instead of MyBatis when retrieving connector messages for the message browser/exporter. This provides a performance bonus of roughly 40% for this stage of the retrieval.

MIRTH-2467
Added Errors column to the message browser to replace the Response Sent/Status column
This column indicates whether a Processing, Postprocessor, or Response error occurred in the message (or Multiple). This is necessary because we no longer retrieve the error (or maps) during the initial search.
The ConnectorMessage stores an errorCode during processing which indicates which errors occurred, using bitwise storage.

MIRTH-2436
If the response transformer sets the message status to QUEUED but queueing is disabled, a Response Error will now be stored. This allows both the original Processing Error and the Response Error to be saved and viewed. This particular Response Error is not stored when the status is fixed before the response transformer, because the connectors themselves often set the status to QUEUED unbeknowst to the user.

MIRTH-2462
Added tooltips to the Message Browser columns.
Refactored the columns to make adding/modifying them easier in the future.
Added minimum width on Connector column

MIRTH-2466
Added Orig. Received Date to the message browser, using the date from the message table.

Misc.
Re-added toString method in Response object as it is needed to display a value in the message browser maps tab.
Flipping the boolean returned in FilterTransformerResult now that the logic has been reversed. (Except for the ones that were already previously flipped)
Ran the formatter.
Issue: MIRTH-2426


git-svn-id: https://svn.mirthcorp.com/connect/trunk@6257 c801cdd1-710f-0410-a44a-8d9edceeb29b