Made XStream the default serializer used in Donkey.
Moved most of the XStream code that was in ObjectXMLSerializer on the Server into a new XStreamSerializer class in Donkey.

Made the Server's ObjectXMLSerializer extend the new XStreamSerializer class in Donkey.

Changed ObjectXMLSerializer to be a singleton, so that plugins can add annotations that will be recognized everywhere. This also may provide a performance benefit since instantiating XStream can be expensive.
Issue: MIRTH-2424


git-svn-id: https://svn.mirthcorp.com/connect/trunk@6279 c801cdd1-710f-0410-a44a-8d9edceeb29b