Removed default message browser column options from settings panel
Column visibility preferences are now persisted automatically when a user shows or hides a column in the message browser table.

Made the column control (at the top right of the message table) use the same popup menu as right clicking a column. Added Horizontal Scroll to the popup menu.
Issue: MIRTH-2490


git-svn-id: https://svn.mirthcorp.com/connect/trunk@6364 c801cdd1-710f-0410-a44a-8d9edceeb29b