MIRTH-1700, MIRTH-2552: Updated the migration controller to call custom classes/code to handle server/schema migration, instead of just executing each delta script. The concept of a separate "schema version" is now gone; now SCHEMA_INFO.VERSION stores the product version (e.g. 3.0.0). Instead of simply iterating through the schema versions and calling each delta script, the migration controller now delegates that to custom Java classes (for now, just Migrate2_0_0 and Migrate3_0_0) that can call their respective delta scripts -and- do whatever other custom code may be necessary.
The channel table has been changed so that the entire serialized channel is stored in a single column. The ID, name, and revision columns remain, however. This is an example of where "custom code" is needed on server migration. Migrate3_0_0 executes all the delta scripts up to the 3.0.0 script (which renames the old channel table and creates the new one). Then it manually selects data out from the old table, builds up the 2.x model XML as a document, and stores the serialized data in the new table.

The migrateContents and migrateServerProperties functions that were in the migration controller have been moved to Migrate2_0_0, since the upgrade from 1.8.2 to 2.0.0 is the only place they were necessary anyway.

git-svn-id: https://svn.mirthcorp.com/connect/trunk@6458 c801cdd1-710f-0410-a44a-8d9edceeb29b