Fixed outdated channel being deployed with no warning in certain cases if a channel is re-imported and deployed immediately after it is deleted.
Modifed updateChannel to use getCachedChannelById instead of getChannel to determine whether a channel already exists. This implicitly updates the channel cache before a channel is saved.

Synchronized updateChannel and removeChannel in DefaultChannelController to ensure that the channel cache and database can never have different versions of a channel at the same time.
Issue: MIRTH-2642


git-svn-id: https://svn.mirthcorp.com/connect/branches/2.x@6508 c801cdd1-710f-0410-a44a-8d9edceeb29b