Added the ability to specify a Migrator class for plugins, which is invoked on startup to run any migration code for the plugin.
Added the ability to migrate private connectors to 3.0.0 by using reflection to look for a 'migrate3_0_0' method on the connector properties class in the private connector plugin. Made several of the utility methods in ImportConverter3_0_0 public so that private connectors can use them in their migration code.

Renamed DatabaseSchemaMigrationException to MigrationException

Renamed the previous ServerMigrator class to 'Migrator'.

Moved code in DefaultMigrationController that performs the schema initialization and migration into a new ServerMigrator class that extends the Migrator abstract class. This was done so that the core server migration will follow the same pattern as plugins that define a single Migrator class that manages the schema version and runs the appropriate migration code.
Issue: MIRTH-2647


git-svn-id: https://svn.mirthcorp.com/connect/trunk@6511 c801cdd1-710f-0410-a44a-8d9edceeb29b