MIRTH-2682
Fixed problem with viewing messages in the message browser that contain invalid XML character(s).

Modified ObjectXMLSerializer so that it only puts the serialized data into a DOM document if it detects that the serialized data contains a Migratable object.

More work is needed to ensure that invalid XML characters are dealt with appropriately in all cases (see MIRTH-2692).

git-svn-id: https://svn.mirthcorp.com/connect/trunk@6570 c801cdd1-710f-0410-a44a-8d9edceeb29b