MIRTH-2699: Fixed bug that caused the StartTask rollback to not set the state of the individual connectors. Now, each source and destination connector have their own rollback mechanisms in onStop() and onHalt().
In SourceConnector.stop() if an error occurred in onStop() and the underlying cause is not an InterruptedException, the current state is always set to STOPPED. In SourceConnector.halt(), the current state is always set to STOPPED regardless of whether an error occurred. The same changes have been made for DestinationConnector.

Also, a change has been made to StartTask so that in the rollback if an exception happened while stopping, the channel attempts to halt the connector(s) instead.

git-svn-id: https://svn.mirthcorp.com/connect/trunk@6581 c801cdd1-710f-0410-a44a-8d9edceeb29b