The global postprocessor no longer returns anything be default. By default the channel postprocessor's return value would be used.
Default global postprocessor scripts that have been migrated from older versions should no longer cause the global postprocessor to run.

If the global postprocessor returns nothing, the response from the channel postprocessor should be used.

If the global postprocessor returns a response, that response should be used.

If the global postprocessor returns something other than a response, a new response will be created with that object's string representation as the message.
Issue: MIRTH-2708


git-svn-id: https://svn.mirthcorp.com/connect/trunk@6593 c801cdd1-710f-0410-a44a-8d9edceeb29b