Made the borders of the variable list and the file pattern components line up by not using a titled border for the variable list.
Increased the default height of the message export dialog and the message archiver panel so that all the variables in the variable list are visible and so there is room for future additions.

Added a note at the bottom of the panel communicating to the user that attachments are not included with archived/exported messages.
Issue: MIRTH-2750


git-svn-id: https://svn.mirthcorp.com/connect/trunk@6682 c801cdd1-710f-0410-a44a-8d9edceeb29b