MIRTH-2447 MIRTH-2768
Rewrote the reconnect logic for the JMS connectors. Fixed bug that resulted in a new reconnect thread being created each time a message attempted to send from the dispatcher.

Also changed behavior so that if a message fails to send due to a connection error, it will attempt to reconnect before allowing the message to error.

Updated tooltip on the retry interval field in the JMS Writer panel to communicate that the interval is only relevant when the connector is idle.

git-svn-id: https://svn.mirthcorp.com/connect/trunk@6693 c801cdd1-710f-0410-a44a-8d9edceeb29b