MIRTH-2804 Removed unused indexes.
Determined that Postgres does not use the index on (message_id, id) even when those are the only columns in the where clause. It uses the (messageid, id, status) index instead.

MIRTH-2792
Added new index to the d_m table to account for queries that now must scan for both id and server_id.

git-svn-id: https://svn.mirthcorp.com/connect/trunk@6707 c801cdd1-710f-0410-a44a-8d9edceeb29b