Destination Connector now has a method getDispatcherId() which provides an ID to a dispatcher that can be used to identify dispatcher specific variables. If the thread calling it is not a queue thread, it will return -1. If the thread is a queue thread and there is only 1 queue thread, it will return -1. Otherwise it will return the queue thread's id.
Destination queue threads are now stored in a map with the thread's id as a key. The map is now cleared when the connector is stopped or halted.

The TcpDispatcher, WebServiceDispatcher, HttpDispatcher, and DatabaseDispatcher are now thread safe by utilizing dispatcher id.

DatabaseDispatcherQuery now cleans up its data sources on stop/halt instead of on undeploy.

ValueReplacer now uses an AtomicLong to keep track of the count.
Issue: MIRTH-2791


git-svn-id: https://svn.mirthcorp.com/connect/trunk@6719 c801cdd1-710f-0410-a44a-8d9edceeb29b