Reordered method signatures in DonkeyDao so that all retrieval methods are below the insert/update/delete methods.
Reordered unit tests in DonkeyDaoTests so that the tests are in the same order as the method signatures in DonkeyDao. Also added TODOs to add unit tests for methods that don't yet have tests.

Ran unit tests for all databases and fixed some minor issues with the tests.
Issue: MIRTH-2250


git-svn-id: https://svn.mirthcorp.com/connect/trunk@6807 c801cdd1-710f-0410-a44a-8d9edceeb29b