Added localChannelId field to donkey's Channel class.
Changed JdbcDao.getLocalChannelId() so that it first attempts to look up the localChannelId for a given channelId from Donkey's list of deployed channels. If not found, then it queries the d_channels table. This is done so that the d_channels table is not queried each time a message is processed.
Issue: MIRTH-2861


git-svn-id: https://svn.mirthcorp.com/connect/trunk@6812 c801cdd1-710f-0410-a44a-8d9edceeb29b