When importing a message with received, queued, or pending status, set the status to ERROR so recovery and queues do not pick them up.
Set processed to false when importing a message.

Don't update stats when inserting an imported connector message.

Don't decrement stats when reprocessing and replacing a message that was imported.

Set Import ID and Import Channel Id to null if a message is reprocessed and replaced.

Added separator to message import and export dialog and packing them instead of setting a fixed size.
Issue: MIRTH-2910


git-svn-id: https://svn.mirthcorp.com/connect/trunk@6871 c801cdd1-710f-0410-a44a-8d9edceeb29b