Overloaded SourceConnector.dispatchRawMessage to take a parameter that will force the message to be processed even if the connector is stopped (channel is paused).
DonkeyEngineController.dispatchRawMessage now takes the force parameter and passes it to SourceConnector. Updated related references.
Messages being reprocessed or sent from the administrator will set force to true, which allow the messages to be processed even if the channel is paused.

If messages are being reprocessed when a channel is stopping or stopped, the entire batch will be cancelled (but not rolled back). This prevents an error message for occurring for each new reprocessed message.

MIRTH-2925
MIRTH-2926

git-svn-id: https://svn.mirthcorp.com/connect/trunk@6891 c801cdd1-710f-0410-a44a-8d9edceeb29b