Fixed errors when attempting to export to or import from a relative path on the filesystem.
Importing/exporting from/to the server (and the archiver) uses the directory that the server was launched from to resolve relative paths. Importing/exporting from/to the client uses the user home directory to resolve relative paths.

Importing/exporting using the CLI will use the current working directory to resolve relative paths.

Updated the message import methods to return a MessageImportResult object instead of int[], for better readability.
Issue: MIRTH-2912


git-svn-id: https://svn.mirthcorp.com/connect/trunk@6897 c801cdd1-710f-0410-a44a-8d9edceeb29b