MIRTH-2945
Reverted the previous commit related to this JIRA issue that made the LoadedExtensions.dashboardPanelPlugins field a LinkedHashMap, since that had an unintended side-effect with the ordering of the dashboard plugin tabs.

Instead, a setInitialDashboardTabComponent() method has been added to Frame, that will allow a dashboard plugin to set itself as the initially displayed tab.

This feature in the code is temporary and will be superceded by MIRTH-3034, in which the plugin with the greatest priority/weight will be the first one displayed.

git-svn-id: https://svn.mirthcorp.com/connect/trunk@6998 c801cdd1-710f-0410-a44a-8d9edceeb29b