MIRTH-2543
Fixed problem with how the WebAdmin connects to the server when authenticating users. Previously, it extracted the server address from the request URL, which could be incorrect if the client's network has routed the server's address/port to a different address or port that is not accessible by the WebAdmin. In that case, the user would be unable to log in via the WebAdmin.

It now determines the server's address from reading mirth.properties, which should resolve the above problem.

git-svn-id: https://svn.mirthcorp.com/connect/trunk@7012 c801cdd1-710f-0410-a44a-8d9edceeb29b