MIRTH-3061
Fixed error when migrating to version 3.0.0 when using a MySQL 5.0.x or 5.1.x database.

There was a bug in these versions of MySQL (http://bugs.mysql.com/bug.php?id=54462) that left the previous transaction open when switching from autocommit off to autocommit on.

As a workaround, added a START TRANSACTION statement to the beginning of the mysql 3.0.0 migration script that will force the previous transaction to close.

Tested with MySQL 5.1.52 and verified that it resolves the migration problem.

Also tested with the latest MySQL 5.6.14 and verified that migration still works with this version as well.

git-svn-id: https://svn.mirthcorp.com/connect/trunk@7024 c801cdd1-710f-0410-a44a-8d9edceeb29b