MIRTH-3064
Fixed potential problem when creating the database schema in Oracle by removing the hard-coded PERSON_ID 1 value when inserting into the PERSON_PASSWORD table. In some cases, a sequence that "starts with 1" will actually start with 2. Replace the hard-coded value with PERSON_SEQUENCE.CURRVAL which will return the correct PERSON_ID that was used to create the default 'admin' user.

git-svn-id: https://svn.mirthcorp.com/connect/trunk@7032 c801cdd1-710f-0410-a44a-8d9edceeb29b