MIRTH-3172: Updated everything to version 3.0.2.
Added migrate3_0_2 method to the Migratable interface, and also to all implementations.
Added call to migrate3_0_2 in MigratableConverter.
Added V3_0_2 to the Version enum.
Added 3.0.2 to the switch statement in ServerMigrator.
Updated all creation SQL scripts to insert version 3.0.2.
Updated mirth-build.properties to pass in 3.0.2.
Updated all plugin/connector metadata files to use 3.0.2 for the mirthVersion.

git-svn-id: https://svn.mirthcorp.com/connect/trunk@7067 c801cdd1-710f-0410-a44a-8d9edceeb29b