MIRTH-3164, MIRTH-3165: Added the ability for users to easily change the max heap size used by the Administrator on the launch page.
There are now two new possible entries in mirth.properties. The first, administrator.maxheapsize, determines the default heap size set in the JNLP served via the web start servlet. The second, administrator.maxheapsizeoptions, determines what heap size settings are available for the user to choose from on the launch page. We add the former setting automatically on migration, but the latter setting is hidden; when absent the default is 256/512/1024/2048 MB.

On the launch page there is now a cog button next to the launch button, which opens up a dropdown menu. In that menu, the user can specify the max heap size for the Administrator. That selection is passed into WebStartServlet in the GET request, and is used to populate the max-heap-size attribute on the j2se node. If not specified in the request, the servlet will instead use the value in mirth.properties. If mirth.properties doesn't have it, it will use our new default of 512 MB, hard-coded in mirth-client.jnlp.

git-svn-id: https://svn.mirthcorp.com/connect/trunk@7079 c801cdd1-710f-0410-a44a-8d9edceeb29b