MIRTH-3150: Fixed function code template inclusion so that it works the same as it did in 2.x. The only differences are that Channel code templates are now available in the channel deploy/shutdown/preprocessor/batch scripts where they weren't available before. The attachment script (new in 3.x) allows Channel or higher, and the response transformer (also new) allows Message or higher (i.e. all templates).
In the global scripts panel, we're now actually restricting the available templates shown in the UI to only Global Channel or Global. This is because although Channel templates are available for the postprocessor, they aren't available for the other scripts, and the panel doesn't differentiate between different script types when showing the reference list.

git-svn-id: https://svn.mirthcorp.com/connect/trunk@7125 c801cdd1-710f-0410-a44a-8d9edceeb29b