MIRTH-3264, MIRTH-1985, MIRTH-3265: Upgraded the HTTP Sender and ServerConnection to use HttpComponents rather than Apache HttpClient.
MIRTH-2946: The response timeout is now being used to set the socket connect timeout as well as the SO_TIMEOUT value.

MIRTH-3293: The HTTP Listener is now correctly reading GZIP input streams. Before the input stream was converted to a string first using the request charset, but that was incorrect if the content was GZIP-encoded. Now it's checking for the correct header and wrapping the input stream in a GZIPInputStream before attempting to read from it.

git-svn-id: https://svn.mirthcorp.com/connect/trunk@7174 c801cdd1-710f-0410-a44a-8d9edceeb29b