MIRTH-3365 Added batch support for connectors other than File Reader and TCP Listener.
MIRTH-2132
Added an option to check for the last message in a batch which will add a "batchComplete" value in the sourceMap. In order to achieve this, we perform a look ahead for to see if another message exists in the batch before processing the current one. 

MIRTH-3252
Added an option to use either the response from either the first or last message in a batch as the response that is send back to the source system. 

git-svn-id: https://svn.mirthcorp.com/connect/trunk@7217 c801cdd1-710f-0410-a44a-8d9edceeb29b