MIRTH-3365 MIRTH-3352 Renamed Response Settings to Source Settings and Moved the Process Batch and Batch Response options into Source Settings. Whether batch is supported is now determined by the connector via the canBatch method on SourceConnectorPropertiesInterface.
Batch Response options are First and Last, which indicates whether the response from the first or last message in the batch will be used. 

Added migration code for batch properties.

Updated method for bolding group headers on the DataTypePropertiesPanel so it will work even for a group with no properties.

git-svn-id: https://svn.mirthcorp.com/connect/trunk@7237 c801cdd1-710f-0410-a44a-8d9edceeb29b