MIRTH-3382
Created a DataPrunerController abstract class along with a DefaultDataPrunerController implementation. DataPrunerService now delegates commands to DataPrunerController. The previous logic in DataPrunerService has been moved into DefaultDataPrunerController.

Added a temporary hard-coded check in DataPrunerController.getInstance() to use a ClusterDataPrunerController implementation if it is available. This will be removed when MIRTH-3351 is completed.

Also removed the pruner isRunning check when the stop method is invoked since the running boolean is already checked in the DataPruner.stop() method.

git-svn-id: https://svn.mirthcorp.com/connect/trunk@7239 c801cdd1-710f-0410-a44a-8d9edceeb29b