MIRTH-2181, MIRTH-3323: Both the HTTP Listener (for responses) and the HTTP Sender now have the ability to send binary content. The listener now has a "Response Data Type" option that automatically Base64-encodes data going back to the source. The sender also has a "Data Type" option that works essentially the same as it does for TCP and File.
Also, both the receiver and dispatcher have been modified to better handle incoming binary data. When the receiver gets a request with a binary content type, it will automatically convert that to a byte array and store that in the RawMessage instead (assuming Plain Body is enabled). When the sender gets a response with a binary content type, it will do the same thing, so it's now possible to send and receive binary content without doing crazy charset hacks like windows-1256 stuff.

git-svn-id: https://svn.mirthcorp.com/connect/trunk@7252 c801cdd1-710f-0410-a44a-8d9edceeb29b