MIRTH-3387 A new reference has been added to the JavaScript scope: "destinationSet". This represents all the destinations that will the message will be sent to. Users can remove destinations using the provided methods on the DestinationSet object: remove, removeAllExcept, removeAll.
The constant mirth_destinations has been renamed to destinationSet. The sourceMap value for this constant is now read only. 

JavaScript executors now take in destinationIdMap instead of destinationNameMap, which is the same thing without the "d" prepended to the metaDataId. ResponseMap prepends the "d" directly to account for this change since it was the only place it destinationNameMap was used.

MIRTH-3391
MIRTH-3392
Fixed setting destinationMetaDataIds on RawMessage from JavaScript not working
destinationMetaDataIds can now be a JavaScript Array or any Java Collection

git-svn-id: https://svn.mirthcorp.com/connect/trunk@7273 c801cdd1-710f-0410-a44a-8d9edceeb29b