Messages are now logged on channel startup if and only if messages are being recovered for a channel.
This is done by setting a boolean flag to determine if a logging message has already occurred and checking whether or not the dao returns messages to be recovered. If logging has not occurred and there are messages to recover, a log statement will be sent to the dashboard.

If at the end of the process the boolean flag was found to be true, a message will be written to the dashboard stating that the process has completed.

log4j properties were also modified to reflect info logging for the RecoveryTask
Issue: MIRTH-3179


git-svn-id: https://svn.mirthcorp.com/connect/trunk@7299 c801cdd1-710f-0410-a44a-8d9edceeb29b