Added ability to inject map variables in the channel writer.
A metadata variable table was added to the channel writer which allows the user to input map keys, and a list was added to the VmDispatcherProperties to store the values from the table. The VmDispatcher attempts to replace each key by doing an ordered lookup across all maps. If the value is found, it is injected into the source map, otherwise it is ignored.
Issue: MIRTH-3380


git-svn-id: https://svn.mirthcorp.com/connect/trunk@7301 c801cdd1-710f-0410-a44a-8d9edceeb29b