MIRTH-3433
Fixed deadlocks in SQL Server when multiple servers are updating message statistics for a channel.

The updateChannelStatistics and updateConnectorStatistics sql statements are now overridden for SQL Server in order to supply the TABLOCK hint to the query optimizer. This forces a table lock to be acquired when updating the message statistics, which prevents deadlocks from occurring in a shared database configuration.

git-svn-id: https://svn.mirthcorp.com/connect/trunk@7379 c801cdd1-710f-0410-a44a-8d9edceeb29b