MIRTH-3475
In Donkey's ChannelController, moved the retry logic added to createChannel() into the getLocalChannelId() method, since the server may also need to re-fetch the list of localChannelIds in the even that two servers are attempting to create donkey tables for the same channel.

Also added a 500 millisecond sleep in between retries to allow time for the server creating the donkey channel tables to finish.

git-svn-id: https://svn.mirthcorp.com/connect/trunk@7404 c801cdd1-710f-0410-a44a-8d9edceeb29b