Now validating on the server that the username is unique when updating a user. This was only previously being done when creating a new user.
When the server fails validation for unique username, the Administrator is now properly handling the exception and showing a warning. The Administrator will also return false to indicate the user was not saved, instead of just continuing and doing bad things.

Updated the warning message when a username already exists.
Issue: MIRTH-3490


git-svn-id: https://svn.mirthcorp.com/connect/trunk@7420 c801cdd1-710f-0410-a44a-8d9edceeb29b