MIRTH-3492: Split up the protocol list in mirth.properties into a client-specific list and server-specific list. This is because by default, Java enables a different set of protocols on the client side than it does on the server side.
In Java 7, TLSv1.1, TLSv1.2, and SSLv2Hello were not enabled by default on the client. According to the JSSE provider specifications, this is because some server don't implement forward-compatibility and cannot accept the newer formats. The SSLv2Hello protocol is only a pseudo protocol wrapper, but is also disabled by default because some newer servers choose to no longer accept the old ClientHello format.

In Java 8, TLSv1.1 and TLSv1.2 are enabled by default on the client, but SSLv2Hello is not.

Because of these discrepancies and because we want to allow for workarounds in case users are actually on older versions of Java, we split up the protocols list so users can override the defaults for the client and server separately.

git-svn-id: https://svn.mirthcorp.com/connect/trunk@7458 c801cdd1-710f-0410-a44a-8d9edceeb29b