The document writer now has the option to output the content into an attachment, and as such, an option was added to the DocumentDispatcherProperties to determine whether or not an attachment should be created based off of the new GUI options added to the DocumentWriter.
The DocumentDispatcher will now always write the contents to a ByteArrayOutputStream. If a file option was selected, these bytes will then get written to an output file.
If an attachment option was selected, an attachment will get created from these bytes and stored with the message. If an attachment was generated, the Id will be stored in the response message.

The Base64Util was also updated to allow the creation of unchunked byte arrays.
Issue: MIRTH-2389


git-svn-id: https://svn.mirthcorp.com/connect/trunk@7491 c801cdd1-710f-0410-a44a-8d9edceeb29b