Attachments are now able to exported and/or be archived within Mirth Connect. The import feature has also had its functionality extended to be able to import attachments along with messages.
UI options have been added to provide the additional attachment functionality, and the attachment label has been removed since it's no longer relevant.

An AttachmentSource interface was added to the server project which is responsible for returning a list of attachments. This list of attachments, if not empty, will now be added just before writing out the message via the MessageExporter.

The MessageExporter now takes in a reference to the AttachmentSource. If the source is not-null, attachments will be added to the message given the implementation provided.

The Channel class will now insert message attachments into the database where applicable. Since we are now allowing attachments to be imported back in, queries to return an attachment (used by the viewer) had to be modified to ensure the correct attachment is sent back to the viewer. All code references have been updated to account for this including modifying the original query to include a reference to the message Id.

MIRTH-2423
MIRTH-3516


git-svn-id: https://svn.mirthcorp.com/connect/trunk@7504 c801cdd1-710f-0410-a44a-8d9edceeb29b