A new block size field is available on the Data Pruner settings panel, with a default size of  100 unless otherwise specified by the user.
The validation method within the MessageExportPanel was refactored to return a string instead of a boolean value. This was done to give specific details as to what errors might've occurred within the panel.

The tooltip for including attachments during export/archiving was also ninja hot fixed since it previously did not exist.
Issue: MIRTH-3531


git-svn-id: https://svn.mirthcorp.com/connect/trunk@7514 c801cdd1-710f-0410-a44a-8d9edceeb29b