MIRTH-796 MIRTH-3561
Added the ability to select a user when creating an alert action, which will send an email to the user's email address (if one is defined).

Re-designed alert actions so that additional types of alert actions can be defined by plugins or any external code.

Heavily refactored the AlertActionPane class to accomodate the above change.

Added two new client API methods:

1) getAlertInfo: obtains the alert model, alert protocols and protocol options and also channel summaries. Used by AlertEditPanel when editing/creating an alert.

2) getAlertProtocolOptions: obtains alert protocols and their corresponding lists of options. Used by the AlertActionPane on initialization.

git-svn-id: https://svn.mirthcorp.com/connect/trunk@7515 c801cdd1-710f-0410-a44a-8d9edceeb29b