The file directory, filter pattern, username and password are now being replaced during each poll.
The connectorProperties argument was removed for the getConnection method in lieu of passing in the username and password directly since these were the only properties used. All code that referenced this method has been updated to reflect these changes.

Replacement is now done onStart instead of on deploy and also once per poll. The readDir is now local to the polling method since that's the only time it's used.
Issue: MIRTH-3131


git-svn-id: https://svn.mirthcorp.com/connect/trunk@7518 c801cdd1-710f-0410-a44a-8d9edceeb29b