MIRTH-3568
Fixed exception handling in the data pruner so that manually stopping the data pruner no longer generates an error in the server log.

Also fixed a couple bugs that prevented the 'Data Pruner Halted' server event from being dispatched. This code was never run previously because the InterruptedException generated by the manual halt was not being caught correctly.

git-svn-id: https://svn.mirthcorp.com/connect/trunk@7529 c801cdd1-710f-0410-a44a-8d9edceeb29b