MIRTH-1461 The MirthTable now has a Preference object to persist the sorted order of columns and uses a prefix, along with a designated key, in order to retrieve the associated settings. An additional constructor has been added that takes in this prefix and instantiates the preferences. The sorterChanged event was overridden to now serialize the column order if a prefix has been provided (the prefix assumes that the changes should be persisted). The setModel function was also overridden so as to allow the persisted settings to be restored, and finally a restoreDefaults method was implemented to reset the settings back to their original state.
The ChannelPanel's channelTable is now created using a prefix. The stipulation to sort columns by via the name has been removed to allow the above functionality. The exposed restoreDefault method is now called out within the "Restore Default" menu item.

The MirthTreeTable's implementation to provide the same functionality differs slightly in that it instead uses methods exposed by the SortableTreeTable. An additional constructor was also added which takes in a prefix, determining whether or not settings should be persisted. A method has been created, saveSortOrder, which takes in the index of the sorted column, persisting the index and the actual sort order retrieved from the parent class. Finally, the class now has a public method to restore the settings to their original state.

The DashboardPanel now has a right click context menu for the column headers. This was added to provide a way to restore defaults and clear out the user preferences. Each time the columns sorted, the statusTable saves the sorted order. When the table is built for the first time, the persisted settings are deserialized and applied to the corresponding column.
Issue: MIRTH-1461


git-svn-id: https://svn.mirthcorp.com/connect/trunk@7532 c801cdd1-710f-0410-a44a-8d9edceeb29b