MIRTH-3574 Moved the interrupted check outside of runDeleteQueries/runDelete so that it is not possible to interrupt the deletes for a given range. This ensures that messages are fully deleted before the pruner stops to avoid leaving any orphaned records for SQL Server and Derby
MIRTH-3575
Only add the message Id to contentMessageIds if it is not added to messageIds. This way content is not pruned independently if its going to be pruned with the whole message anyway.

MIRTH-3576
Swapped the order that messages and content are pruned so that whole messages are pruned first. This ensures that usually the oldest messages will be pruned first.

git-svn-id: https://svn.mirthcorp.com/connect/trunk@7536 c801cdd1-710f-0410-a44a-8d9edceeb29b