MIRTH-3024 The MirthTreeTable will now persist the order of columns in addition to saving its sorted order.
The table has a mouseListener registered on the tableHeader which is responsible for saving the sorted order of columns. Except for any metaDataColumns (applicable only to the MessageBrowser), the column order is saved in a map, serialized as XML and stored within a preference node. If a new column is added to the panel, the column ordering will be reset to its default order, otherwise column ordering will be restored from this map.

The MessageBrowser has been updated to now persist the order of columns. MetaDataColumn preferences will only persist across the same session. Logging out and back in will reset these columns to their default states.

The DashBoard panel has also been updated to persist column ordering since the underlying component is the same MirthTreeTable.
Issue: MIRTH-3024


git-svn-id: https://svn.mirthcorp.com/connect/trunk@7537 c801cdd1-710f-0410-a44a-8d9edceeb29b