MIRTH-3570
Upgrade the commons-dbcp and commons-pool libraries to the latest versions of commons-dbcp2 and commons-pool2 respectively.

Updated the DBCPConnectionPool class to use the new dbcp2 api for constructing the connection pool.

Updated the Database writer connector to use dbcp2.

Updated the File reader/writer connectors to account for api changes in commons-pool2.

Updated DocumentWriterService so that it no longer directly references FileSystemConnection from the file connector, which would have been affected by these changes.

git-svn-id: https://svn.mirthcorp.com/connect/trunk@7590 c801cdd1-710f-0410-a44a-8d9edceeb29b