MIRTH-3570
Modified the Database Writer (DatabaseDispatcherQuery) to no longer use BasicDataSource from DBCP, since it still attempts to invoke Connection.isValid, which fails when connecting to SQL Server if no validation query is supplied.

Replaced BasicDataSource with a custom SimpleDataSource class that persists a single connection and recreates the connection if any of the connection properties change or if the connection has been idle for more than 5 minutes.

Removed the retry feature since it relied on a Connection.isValid check to determine if a retry was needed due to the connection becoming stale. Since we now enforce an idle timeout, this should no longer be necessary.

Removed the dbcp2 library from the Server project since it is no longer used by any Server code.

git-svn-id: https://svn.mirthcorp.com/connect/trunk@7605 c801cdd1-710f-0410-a44a-8d9edceeb29b