MIRTH-3639
Updated 'getUserCredentials' and 'getLatestUserCredentials' queries in Postgres, Oracle and Derby to ORDER BY PASSWORD_DATE DESC NULLS LAST.

This fixes the issue in which passwords with a null password_date are always recognized as the newest password.

Also updated all databases so that passwords with a null password_date are pruned by the 'prunePasswords' query.

git-svn-id: https://svn.mirthcorp.com/connect/trunk@7625 c801cdd1-710f-0410-a44a-8d9edceeb29b