Merged from trunk.
MIRTH-3657: Objects from custom resources placed in variable maps will now be serialized/deserialized with a serializer instance that has the correct resource libraries on the classpath. There's a new SerializerProvider interface that Donkey uses to retrieve a serializer for a specific metadata ID. The interface is passed into BufferedDaoFactory, which gets set when the channel is deployed. Even if the resource gets reloaded, the provider interface will always get the correct serializer from the correct context factory because it's going directly to the context factory controller.

git-svn-id: https://svn.mirthcorp.com/connect/branches/3.2.x@7645 c801cdd1-710f-0410-a44a-8d9edceeb29b