MIRTH-3668 Table components now have been updated to include column controls. Column preferences are now being persisted in a way that makes more sense and will no longer cause the UI to lock up under certain instances, instead failing gracefully.
Both MirthTable and MirthTreeTable now load the preferences from within the constructor and into an in-memory reference. Both tables now also provide a MouseAdapter which will correctly register sorting actions on the header when setMirthColumnControlEnabled is set to true. The MirthTreeTable contains an additional MouseAdapter to control sorting actions since there is no native Java solution. When saving columns, three loops are now performed which will enable all column visibility, move columns to their intended position, and finally disable columns that should not be shown. This is done so that we can keep track of all column orders and column identifiers which dictates shown behavior when a new column is added. When restoring columns, indices are now read into a TreeMap which restores view indices based on the sorted order of the map.

The SortableHeaderCellRenderer now validates the model index since the table is initially built with dummy columns that will throw an invalid index exception.

The SortableTreeTable now correctly compares the model index using the SortableTreeTableModel as opposed to the underlying JXTreeTable model which often differs in contents.

The ChannelPanel, DashboardPanel, EventBrowser and MessageBrowser have had their references updated to reflect the new changes.
Issue: MIRTH-3668


git-svn-id: https://svn.mirthcorp.com/connect/trunk@7671 c801cdd1-710f-0410-a44a-8d9edceeb29b