MIRTH-2288
For some reason this class didn't commit.
Issue: MIRTH-2288

A new task has been added to the ChannelPanel to view messages and only shows up if a single channel is selected. Messages viewed from this context, when the channel is not deployed, have limited options to remove messages.

The showMessages task now defers to either the dashboardPanel or channelPanel to get the channelId for the selected channel. If the current panel is the Channels view, then the deployed state is determined from the list of deployed channels. If the channel is not in this list, then the channel is not deployed.

The removeAllMessages task has also been refactored to defer to the correct panel. If the channelPanel has been viewed and the reference is non null, the selectedChannles are retrieved, and if this reference is not empty, that is to say a channel is selected, the channelId is added to the set of channelIds to remove messages from and the deployed state is checked. Otherwise, the task builds the channel list and checks the deployed state as per usual.

The MessageBrowser now disables any reprocessing message tasks when the channel is not deployed.

The message controller now


git-svn-id: https://svn.mirthcorp.com/connect/trunk@7699 c801cdd1-710f-0410-a44a-8d9edceeb29b