MIRTH-1905 A new alertConflict method has been added to the frame class which creates a new dialog, given a message, to display the number of conflicts with a checkbox to apply the option to all conflicts.
An enum has been created which will now pass back the result of the above selection.

The exportChannels method now performs an initial check to get the number of file collisions. It then defers to the aforementioned method within a loop to determine user action and perform the necessary exports.
Issue: MIRTH-1905


git-svn-id: https://svn.mirthcorp.com/connect/trunk@7710 c801cdd1-710f-0410-a44a-8d9edceeb29b