MIRTH-841 PollingConnectors have now been improved to create schedules which run from the hour as opposed to from start/deploy time. Users can now exclude/include certain weekdays or define a discrete day of month and define active time ranges when using an interval. The user now also has the option to schedule polling to occur at a specific time of the day also on certain weekdays or on a particular day of the month. If none of these options fit the users needs, there is now a "Cron" option which will run schedules based on any defined, valid expressions.
The PollConnector has been changed to take advantage of the Quartz library, and the PollConnectorTask has also been converted from a TimerTask to an InterruptableJob. All of the previous operations were migrated to use the Quartz equivalence to keep the functionality in tact.

There is now a PollConnectorJobHandler class which is used to configure the JobDetail and Scheduler that's used in the actual PollConnector. The PollConnectorJobHandler also has a reference to the new TriggerFactory class whose purpose it is to create job triggers.

The TriggerFactory takes a set of properties and will build the appropriate schedules and return the Trigger associated with the properties, either a daily interval trigger, a daily time trigger or a list of cron triggers.

The new PollConnectorJobFactory class has been created to facilitate in the creation of new jobs, taking in a reference to the current polling connector. This factory will return a PollConnectorJob which has the express purpose of calling poll on the PollConnector. A DummyJob has also been added so that the client may create dummy triggers to calculate next-scheduled polling times.

A new CronProperty, PollingType and PollConnectorPropertiesAdvanced model were added to the Donkey project to store additional properties used in creating Triggers, and the PollConnectorProperties now has a reference to these advanced properties.

On the client side, the PollingSettingsPanel has been reworked and created in MigLayout. The three scheduling types will now dynamically hide/show the appropriate properties based on the selection within the combobox. When interval/time is selected and the advanced properties button clicked, a new AdvancedPollingSettingsDialog will appear which stores the weekly, day of month and active time configurations.
Issue: MIRTH-841


git-svn-id: https://svn.mirthcorp.com/connect/trunk@7733 c801cdd1-710f-0410-a44a-8d9edceeb29b