MIRTH-841 The new polling settings have been applied to the Data Pruner.
The AdvancedPollingSettingsDialog, PollingSettingsPanel and DataPrunerPanels have been modified to accommodate the above. The DataPrunerPanel is now written in MigLayout and has been modified to account for the new polling properties.

The PollConnectorJobHandler now takes in a boolean to flag that dictates if the handler is being used within a connector. The configureJob now takes in an identity and the scheduleJob method has been modified to allow starting the scheduler. The getNextFireTime method now performs different checks based on the trigger type so as to return an accurate time.

The DefaultDataPrunerController now delegates to a PollConnectorJobHandler when configuring the scheduler and any triggers. All relevant code has been updated to take advantage of this code.

Data pruner properties have been added to migration and server restoration blocks.
Issue: MIRTH-841


git-svn-id: https://svn.mirthcorp.com/connect/trunk@7755 c801cdd1-710f-0410-a44a-8d9edceeb29b