MIRTH-3798, MIRTH-3799: Completely overhauled the HTTP API for Mirth Connect.
Instead of using POST requests with form-urlencoded everywhere, the API is a lot more RESTful and intuitive. The framework is built on JAX-RS (using Jersey), and each endpoint is defined in a core "servlet interface". The interface is accessible by both client and server, and provides all the necessary information about the endpoint, including JAX-RS info and also Mirth-specific info like operations and permissions.

The server side has servlets that implement the interfaces, and that all extend from MirthServlet which handles authorization, auditing, and other common tasks.

The client side uses synthetic interface implementations built with the Java reflect API and invocation handlers. So the caller of an interface method doesn't and shouldn't have to know anything about what HTTP method/headers/params are used in the background. All of that is inferred dynamically and sent into ServerConnection which applies it to the HttpComponents request.

All methods are hosted at /api, or /api/{version}. Swagger UI is also hosted by the web server, so that users can browse and test out the API in a browser. It's built to be fully backward-compatible as well. So even when newer versions are released, users can still hit endpoints at /api/3.4.0 for example.

All extensions have been updated to conform with the new structure. There are no more connector services, no more invokePluginMethod calls, etc. All of that has been standardized on this new servlet interface framework.

git-svn-id: https://svn.mirthcorp.com/connect/trunk@7862 c801cdd1-710f-0410-a44a-8d9edceeb29b