MIRTH-2349: Added the ability for channels to be dependent on other channels for deploy/start tasks.
Dependencies are stored as simply pairs of channel IDs (dependent ID and dependency ID). To calculate dependencies, a directed acyclic graph data structure is constructed from the set of channel ID pairs. From there the sink nodes (that no other node is dependent on) are kept track of, and the graph can be traversed in any direction.

Dependencies are edited in the same Set Dependencies dialog that code template libraries and resources are set. Users can add direct dependencies or dependents to any channel, and can see dependency trees in either direction.

When a deploy/undeploy/start/stop/pause/resume task is performed, any other required channels are calculated, and the user is prompted to ask whether the other channels should be included or not. The user is shown an ordered list of channels and what order they will be handled in.

git-svn-id: https://svn.mirthcorp.com/connect/trunk@7949 c801cdd1-710f-0410-a44a-8d9edceeb29b