MIRTH-3904: Added configuration plugin points for DICOM, SMTP, and TCP connectors. A few new methods were added to the connector / connector settings panels to facilitate that. Mainly it entailed taking some of the configuration being done in the receivers/dispatchers and putting them in a default configuration class instead. The configuration classes implement an interface that is instantiated on connector deploy.
MIRTH-3905: Added API endpoint to the configuration servlet to return server-side protocol / cipher suite information. This is returned as a map with five entries.

MIRTH-3845: Changed the connector plugin properties to be a set, and added a local converter for the class so that invalid properties will be properly handled.

git-svn-id: https://svn.mirthcorp.com/connect/trunk@7954 c801cdd1-710f-0410-a44a-8d9edceeb29b