MIRTH-3917: Fixed bug that caused queued messages to be skipped under specific circumstances.
After deleting queued messages, the queue will get invalidated. When a queue thread is processing a message, there is a short period of time between the point the non-QUEUED status is committed to the database, and when the connector message is released from the queue. If the invalidate (and subsequent size update) is done in this period, then the size will be incorrect after the queue thread releases the connector message.

The destination queue now has a read/write lock to prevent this. When destination queue threads are about to commit a non-QUEUED status to the database, they need to obtain a read lock. Multiple threads can have concurrent read locks.

When the queue is invalidated due to messages being deleted, the caller first obtains the write lock. This will block until all queue threads have released their respective read locks. Then the invalidation and size update is done, without any queue threads being able to commit statuses to the database.

git-svn-id: https://svn.mirthcorp.com/connect/trunk@7990 c801cdd1-710f-0410-a44a-8d9edceeb29b